Kludge for MacOS homebrew
Try to workaround quirk in MacOS where libraries are not installed
in a location where the compiler sees them by default. Use
`pkg-config --libs` and `brew --prefix` to try to find libfl.

Note that libfl is only needed because it saved me time and effort
writing a generic main() routine for flex. If these attempts to get
libfl to link end up being more of a hassle, it may make sense to just
ditch the libfl library.