Add valid line pointers to appease VirtualT emulator
Genuine Model T hardware does not care about the line pointers in a
.BA file as they are regenerated upon loading. However, Virtual T
refuses to load such files.

Line pointers, while not too difficult, were more trouble than all the
rest of tokenization. Each line starts with a pointer to the next
line, but of course we don't know where that will be until we finish
tokenizing the current line.

The solution is to add a function, fixup_ptrs() which rewinds the
output file and puts all the line pointers into the correct spots.