Handle lowercase variable names
M100 BASIC barfs with a syntax error if presented with a lowercase
variable name in a tokenized file.

Fortunately, the fix was easy: since the lexer is already matching all
BASIC keywords plus strings, remarks, and data statements, the only
thing left was variables. So, all we have to do is use
yyput(toupper()) instead of the default ECHO.