Script to create M100 BASIC programs of exact size in bytes
Usage: lorem.sh 5.25 KB > foo.do
       tokenize foo.do foo.ba

(foo.ba is now 5.25 * 1024 == 5376 bytes).

This is useful to see how a tokenizer handles programs that are too
large to fit in memory. For example, theoretically, one could tokenize
a program of exactly 32KB for the Model 100, but the pointers might
wrap around badly if tokenized on a Tandy 200.